refactor: Extract helpers, begin extracting statement tests.
Much of the variation in testing can be handled by validating statement construction. Almost all of it, really, with the notable exceptions of `initialize()` and `moveOn()`.

Anyhow, I intend to extract the remaining statement tests from the `Cmi5` class tests (`cmi5.spec.ts`), and to extend the statement tests to cover more cases.